requirements.txt
pandas~=2.1.2
openpyxl~=3.1.2
googlemaps~=4.10.0
google-cloud-bigquery~=3.13.0
